src/error_correction.py

import numpy as np
from qiskit import QuantumCircuit, Aer, execute
from qiskit.quantum_info import Statevector

class QuantumErrorCorrection:
    def __init__(self):
        self.corrected_states = {}

    def apply_error_correction(self, quantum_circuit: QuantumCircuit):
        """Apply error correction codes to a quantum circuit and return the corrected state"""
        backend = Aer.get_backend('statevector_simulator')
        job = execute(quantum_circuit, backend)
        result = job.result()
        statevector = Statevector(result.get_statevector(quantum_circuit))

        # Apply simple error correction (parity check or other code)
        corrected_state = self.correct_state(statevector)
        return corrected_state

    def correct_state(self, statevector: Statevector):
        """Apply a correction algorithm to a given statevector"""
        # Placeholder for advanced error correction logic
        noise_threshold = 1e-3
        corrected_state = np.where(np.abs(statevector.data) > noise_threshold, statevector.data, 0)
        return Statevector(corrected_state)

    def store_corrected_state(self, key: str, statevector: Statevector):
        """Store the corrected quantum state in memory"""
        self.corrected_states[key] = statevector

    def retrieve_corrected_state(self, key: str):
        """Retrieve a corrected state from memory"""
        return self.corrected_states.get(key, None)

    def clear_corrected_states(self):
        """Clear all stored corrected states"""
        self.corrected_states.clear()
